ChatBot/requirements.txt
accelerate==0.23.0
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
aliyun-log-python-sdk==0.8.6
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.1
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
astunparse==1.6.3
async-timeout==4.0.3
asyncer==0.0.2
attrs==23.1.0
auto-gptq==0.4.2
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
backoff==2.2.1
beautifulsoup4 @ file:///croot/beautifulsoup4-split_1681493039619/work
bidict==0.22.1
blinker==1.6.2
blis==0.7.10
boltons @ file:///croot/boltons_1677628692245/work
boto3==1.28.34
botocore==1.31.34
brotlipy==0.7.0
cachetools==4.2.1
catalogue==2.0.9
certifi @ file:///croot/certifi_1683875369620/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
chainlit==0.7.1
chardet @ file:///home/builder/ci_310/chardet_1640804867535/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
coloredlogs==15.0.1
conda==23.3.1
conda-build==3.24.0
conda-content-trust @ file:///tmp/abs_5952f1c8-355c-4855-ad2e-538535021ba5h26t22e5/croots/recipe/conda-content-trust_1658126371814/work
conda-package-handling @ file:///croot/conda-package-handling_1672865015732/work
conda_package_streaming @ file:///croot/conda-package-streaming_1670508151586/work
confection==0.1.1
contourpy==1.1.0
cryptography @ file:///croot/cryptography_1677533068310/work
cycler==0.11.0
cymem==2.0.7
dataclasses-json==0.5.14
datasets==2.14.5
dateparser==1.1.8
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
Deprecated==1.2.14
dill==0.3.7
dnspython==2.3.0
docopt==0.6.2
drain3==0.9.11
elastic-transport==8.4.0
elasticsearch==8.8.2
exceptiongroup==1.1.1
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
expecttest==0.1.4
faiss-gpu==1.7.2
fastapi==0.100.1
fastapi-socketio==0.0.10
ffmpy==0.3.1
filelock @ file:///croot/filelock_1672387128942/work
filetype==1.2.0
fonttools==4.41.1
frozenlist==1.4.0
fschat==0.2.26
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.31
glob2 @ file:///home/linux1/recipes/ci/glob2_1610991677669/work
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
googleapis-common-protos==1.60.0
gradio==3.42.0
gradio_client==0.5.0
greenlet==2.0.2
grpcio==1.58.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
humanfriendly==10.0
hypothesis==6.75.2
idna @ file:///croot/idna_1666125576474/work
imbalanced-learn==0.11.0
imblearn==0.0
importlib-metadata==6.7.0
importlib-resources==6.0.1
ipython @ file:///croot/ipython_1680701871216/work
isodate==0.6.1
jedi @ file:///tmp/build/80754af9/jedi_1644315229345/work
jieba==0.42.1
Jinja2 @ file:///croot/jinja2_1666908132255/work
jmespath==1.0.1
joblib==1.3.1
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpickle==1.5.1
jsonpointer==2.1
jsonschema==4.17.3
kiwisolver==1.4.4
langchain==0.0.279
langcodes==3.3.0
langsmith==0.0.33
Lazify==0.4.0
libarchive-c @ file:///tmp/build/80754af9/python-libarchive-c_1617780486945/work
lightgbm==3.3.5
lxml==4.9.3
markdown-it-py==3.0.0
markdown2==2.4.10
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mdurl==0.1.2
mkl-fft==1.3.6
mkl-random @ file:///work/mkl/mkl_random_1682950433854/work
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.9
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.1
nh3==0.2.14
nltk==3.8.1
nodeenv==1.8.0
numexpr==2.8.5
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
openai==0.28.0
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-exporter-otlp-proto-http==1.20.0
opentelemetry-instrumentation==0.40b0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
optimum @ git+https://github.com/huggingface/optimum.git@7fc27f6747e692fed45acc583e153d45541bc2e7
ordered-set==4.1.0
orjson==3.9.5
packaging==23.1
pandas==2.0.3
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pathy==0.10.2
peft==0.5.0
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.4.0
pipreqs==0.4.13
pkginfo @ file:///croot/pkginfo_1679431160147/work
pluggy @ file:///tmp/build/80754af9/pluggy_1648024709248/work
preshed==3.0.8
prisma==0.10.0
prompt-toolkit @ file:///croot/prompt-toolkit_1672387306916/work
protobuf==3.20.3
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow==12.0.1
pycosat @ file:///croot/pycosat_1666805502580/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.12
pydantic_core==2.6.3
pydeck==0.8.1b0
pydub==0.25.1
Pygments @ file:///croot/pygments_1683671804183/work
PyJWT==2.8.0
Pympler==1.0.1
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.8.2
python-decouple==3.8
python-docx==0.8.11
python-dotenv==1.0.0
python-engineio==4.7.1
python-etcd==0.4.5
python-graphql-client==0.4.3
python-multipart==0.0.6
python-socketio==5.9.0
pytorch-pretrained-bert==0.6.2
pytz @ file:///croot/pytz_1671697431263/work
pytz-deprecation-shim==0.1.0.post0
PyYAML @ file:///croot/pyyaml_1670514731622/work
rdflib==6.3.2
regex==2023.6.3
requests @ file:///croot/requests_1682607517574/work
rich==13.4.2
rouge==1.0.1
ruamel.yaml @ file:///croot/ruamel.yaml_1666304550667/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1666302247304/work
s3transfer==0.6.2
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.1
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
shortuuid==1.0.11
simple-websocket==0.10.1
simplejson==3.19.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve @ file:///croot/soupsieve_1680518478486/work
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spacy-pkuseg==0.0.32
SQLAlchemy==2.0.20
srsly==2.4.7
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
starlette==0.27.0
streamlit==1.26.0
streamlit-aggrid==0.3.4.post3
streamlit-agraph==0.0.45
streamlit-chatbox==1.1.7
streamlit-option-menu==0.3.6
svgwrite==1.4.3
sympy==1.12
syncer==2.0.3
synonyms==3.20.0
tenacity==8.2.2
thinc==8.1.12
threadpoolctl==3.1.0
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
tomli @ file:///opt/conda/conda-bld/tomli_1657175507142/work
tomlkit==0.12.1
toolz @ file:///croot/toolz_1667464077321/work
torch==2.0.1
torch-geometric==2.3.1
torch-scatter==2.1.1
torch-sparse==0.6.17
torchaudio==2.0.2
torchdata @ file:///__w/_temp/conda_build_env/conda-bld/torchdata_1682362130135/work
torchelastic==0.2.2
torchtext==0.15.2
torchvision==0.15.2
tornado==6.3.2
tqdm @ file:///croot/tqdm_1679561862951/work
traitlets @ file:///croot/traitlets_1671143879854/work
transformers==4.32.1
triton==2.0.0
typer==0.9.0
types-dataclasses==0.6.6
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
uptrace==1.20.0
urllib3 @ file:///croot/urllib3_1680254681959/work
uvicorn==0.23.2
validators==0.20.0
wasabi==1.1.2
watchdog==3.0.0
watchfiles==0.20.0
wavedrom==2.0.3.post3
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
websockets==11.0.3
wrapt==1.15.0
wsproto==1.2.0
xxhash==3.3.0
yarg==0.1.9
yarl==1.9.2
zh-core-web-sm @ file:///workspace/project/trying/2023/chatglm2-6b/zh_core_web_sm-3.6.0-py3-none-any.whl
zipp==3.15.0
zstandard @ file:///croot/zstandard_1677013143055/work
